Reorganised structure and added Carthage support
Now there is one main project that has:
 - The framework target including sources
 - An OS X test target, which only compiles Diffing.swift, no need to
start the simulator for testing anymore
 - The Demo as a subproject. It directly references the three relevant
source files. It could be changed to use Carthage instead, but is kept
without it to demonstrate a minimal effort for getting started
