+++ Demo/DiffyTables WatchKit App/Info.plist
@@ -9,7 +9,7 @@
 	<key>CFBundleExecutable</key>
 	<string>$(EXECUTABLE_NAME)</string>
 	<key>CFBundleIdentifier</key>
-	<string>io.radex.DiffyTables.watchkitapp</string>
+	<string>io.radex.Demo.watchkitapp</string>
 	<key>CFBundleInfoDictionaryVersion</key>
 	<string>6.0</string>
 	<key>CFBundleName</key>
@@ -28,7 +28,7 @@
 		<string>UIInterfaceOrientationPortraitUpsideDown</string>
 	</array>
 	<key>WKCompanionAppBundleIdentifier</key>
-	<string>io.radex.DiffyTables</string>
+	<string>io.radex.Demo</string>
 	<key>WKWatchKitApp</key>
 	<true/>
 </dict>
