+++ Source/Supporting Files/DiffyTables.h
@@ -1,0 +1,19 @@
+//
+//  DiffyTables.h
+//  DiffyTables
+//
+//  Created by Silvan Mosberger on 23/07/16.
+//
+//
+
+#import <WatchKit/WatchKit.h>
+
+//! Project version number for DiffyTables.
+FOUNDATION_EXPORT double DiffyTablesVersionNumber;
+
+//! Project version string for DiffyTables.
+FOUNDATION_EXPORT const unsigned char DiffyTablesVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <DiffyTables/PublicHeader.h>
+
+
